refactor: unify OCI image reference parsing
Amp-Thread-ID: https://ampcode.com/threads/T-019ff0e6-00a6-7028-afbb-9c2ac8ed94aa
