+++ pkg/deploy/imageref/doc.go
@@ -1,0 +1,2 @@
+// Package imageref validates and normalizes OCI image references used by deploys.
+package imageref
